backend/requirements.txt
python-dotenv
requests
pandas
scikit-learn
apache-airflow
dvc
python-dateutil
seaborn
mlflow
flask_cors
flask_pymongo
pytest==7.4.0
pytest-cov==4.1.0